% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_continuous.R
\name{clean_continuous}
\alias{clean_continuous}
\title{Clean Continuous Variable}
\usage{
clean_continuous(var)
}
\arguments{
\item{var}{a continuous variable that you want to check the distribution of.}
}
\value{
Returns how many missing, not applicable values along with what values stop the conversion to numeric. If the variable is able to be converted to a continuous form then returns a histogram plot as well.
}
\description{
Clean Continuous Variable
}
\examples{
clean_continuous(c(rnorm(500),NA,"Not Applicable"))
}
