% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_discrete.R
\name{clean_discrete}
\alias{clean_discrete}
\title{Clean Discrete Variable}
\usage{
clean_discrete(var)
}
\arguments{
\item{var}{discrete variable}
}
\value{
bar chart showing all levels of discrete variable along with number of not applicable and missing values
}
\description{
Clean Discrete Variable
}
\examples{
clean_discrete(sample(c("a","b","c"),100,replace = T))
}
