% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hist.R
\name{clean_histogram}
\alias{clean_histogram}
\title{Cleaned Histogram}
\usage{
clean_histogram(var)
}
\arguments{
\item{var}{a continuous variable that you want to check the distribution of.}
}
\value{
A histogram plot or a list of values
}
\description{
Cleaned Histogram
}
