% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_toc.R
\name{codebook_toc}
\alias{codebook_toc}
\title{Title}
\usage{
codebook_toc(
  names,
  labels,
  low_count = 1,
  up_count = 22,
  add_count = 22,
  want_iter = 1
)
}
\arguments{
\item{want_iter}{}
}
\value{

}
\description{
Title
}
