% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_table.R
\name{cross_table}
\alias{cross_table}
\title{Title}
\usage{
cross_table(
  dataset,
  row,
  col,
  percent,
  notes = NULL,
  caption,
  stratavar = NULL,
  tab_name,
  load_tab = F,
  label_name = "",
  file_path
)
}
\arguments{
\item{row}{variable to put on the row}

\item{col}{variable to put in the columns}

\item{notes}{notes to put in the footer}

\item{caption}{what to caption the table as}

\item{stratavar}{strata variable}

\item{tab_name}{name to save table as}

\item{load_tab}{load an already saved table: T or F}

\item{file_path}{file path to save table}
}
\value{

}
\description{
Title
}
