% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers_exact_result_table.R
\name{fishers_exact_result_table}
\alias{fishers_exact_result_table}
\title{Title}
\usage{
fishers_exact_result_table(
  dataset,
  yvar,
  groupvar,
  load_tab = F,
  tab_name,
  file_path
)
}
\arguments{
\item{file_path}{}
}
\value{

}
\description{
Title
}
