% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_code_finder.R
\name{invalid_code_finder}
\alias{invalid_code_finder}
\title{Title}
\usage{
invalid_code_finder(variable, dataset, write_meta = F, meta_path)
}
\arguments{
\item{variable}{name of variable in dataset to check for invalid codes}

\item{dataset}{}

\item{write_meta}{whether you want to write the invalid codes you find to the meta data file}

\item{meta_path}{this xlsx type meta data file must contain a column labelled possible_invalid_codes and invalid_codes. The possible invalid codes must be separated by a comma and entered as a string}
}
\value{

}
\description{
Title
}
