% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_wal.R
\name{kruskal_wal}
\alias{kruskal_wal}
\title{Title}
\usage{
kruskal_wal(
  dataset,
  yvar,
  groupvar,
  load_tab = F,
  tab_name,
  file_path,
  caption = "",
  outcome_vartype = "categorical"
)
}
\arguments{
\item{caption}{}
}
\value{

}
\description{
Title
}
