% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_interpret_baseline.R
\name{logistic_interpret_baseline}
\alias{logistic_interpret_baseline}
\title{Title}
\usage{
logistic_interpret_baseline(
  variable,
  regression,
  estimate_name,
  chart,
  patient_group,
  categorical = F,
  cat_level = NULL,
  cat_varname = NULL,
  point_est_state = NULL,
  lowerb_state = NULL,
  upperb_state = NULL,
  cat_binary = T
)
}
\arguments{
\item{cat_binary}{}
}
\value{

}
\description{
Title
}
