% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_interpret_main.R
\name{logistic_interpret_main}
\alias{logistic_interpret_main}
\title{Title}
\usage{
logistic_interpret_main(
  variable,
  regression,
  reg_var = NULL,
  reg_level = NULL,
  estimate_name,
  varname_baseline,
  chart = NULL,
  patients_group,
  baseline = "",
  digits = 3,
  categorical = F,
  cat_binary = F,
  cat_level = NULL,
  cat_varname = NULL,
  point_est_state = NULL,
  lowerb_state = NULL,
  upperb_state = NULL,
  conclude_state = NULL,
  benchmark_odds = 0.05
)
}
\arguments{
\item{benchmark_odds}{how far away from one to then consider the odds ratio different from one}
}
\value{

}
\description{
Title
}
