% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_table.R
\name{logistic_reg_table}
\alias{logistic_reg_table}
\title{Title}
\usage{
logistic_reg_table(
  dataset,
  yvar,
  xvars,
  family,
  to_exp = T,
  caption,
  notes = NULL,
  gt_tab = FALSE,
  both_tabs = T,
  load_tab = F,
  tab_name,
  factor_notapp = NULL,
  intercept = F,
  file_path
)
}
\arguments{
\item{yvar}{outcome variable}

\item{xvars}{independent variables}

\item{family}{distribution for outcome variable}

\item{to_exp}{exponentitate the results}

\item{caption}{caption to apply}

\item{notes}{footer notes}

\item{gt_tab}{return the gt table; T or F}

\item{both_tabs}{return both the gt table and the flextable}

\item{load_tab}{load an already saved table}

\item{tab_name}{name to save the table}

\item{factor_notapp}{which variables have not appliable to filter out}

\item{intercept}{show the intercept; T or F}

\item{file_path}{give the file path to save/find the table}
}
\value{

}
\description{
Title
}
