% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_plot.R
\name{rain_plot}
\alias{rain_plot}
\title{Rain Plot}
\usage{
rain_plot(
  dataset,
  varname,
  groupvar = NULL,
  facetvar = NULL,
  title = "",
  y_label = "",
  x_label = "",
  leg.title = "",
  flip = T,
  notes = "",
  wrap_count = 87,
  alpha_val = 0.3,
  boxplot = T,
  ncols = 1,
  color_point_var = NULL
)
}
\arguments{
\item{dataset}{dataset}

\item{varname}{outcome variable}

\item{groupvar}{grouping variable}

\item{facetvar}{facet variable}

\item{title}{title for the plot}

\item{y_label}{label for y axis}

\item{x_label}{label for x axis}

\item{leg.title}{label for the legend}

\item{flip}{TRUE or FALSE flip the x and y axis}

\item{notes}{add notes as a caption at the bottom}

\item{wrap_count}{make the text for the caption fit better}

\item{alpha_val}{control the visibility for the points}

\item{boxplot}{TRUE or FALSE; add}

\item{ncols}{number of columns for the facet}

\item{color_point_var}{what colour should the points be}
}
\value{
a rain plot
}
\description{
Rain Plot
}
