% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch_ttest_plot.R
\name{welch_graph}
\alias{welch_graph}
\title{Title}
\usage{
welch_graph(
  dataset,
  yvar,
  groupvar,
  legend.title,
  y_label = "Frequency",
  x_label = NULL
)
}
\arguments{
\item{x_label}{}
}
\value{

}
\description{
Title
}
