% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_rank.R
\name{wilcox_rank}
\alias{wilcox_rank}
\title{Title}
\usage{
wilcox_rank(
  dataset,
  yvar,
  groupvar,
  load_tab = F,
  tab_name,
  file_path,
  caption = "",
  outcome_vartype = "categorical"
)
}
\arguments{
\item{caption}{}
}
\value{

}
\description{
Title
}
