% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hist.R
\name{clean_histogram}
\alias{clean_histogram}
\title{Cleaned Histogram}
\usage{
clean_histogram(var)
}
\arguments{
\item{var}{a continuous variable that you want to check the distribution of.}
}
\value{
Returns how many missing, not appliable values along with what values stop the conversion to numeric. If the variable is able to be converted to a continuous form then returns a histogram plot as well.
}
\description{
Cleaned Histogram
}
\examples{
clean_histogram(c(rnorm(500),NA,"Not Applicable")
}
